перенёс логику выбора игровой формы и обработчика во вьюху
