+++ js/screens/game.js
@@ -1,8 +1,39 @@
+import {questionTypes} from './../consts';
 import AbstractView from './../view/abstract';
 
 import {statsTemplate} from './../chunks/stats';
-import gameForm from './../chunks/game-form';
+import {gameFormTypeSingle, gameFormTypeSingleHandler} from './../chunks/game-form-single';
+import {gameFormTypeCouple, gameFormTypeCoupleHandler} from './../chunks/game-form-couple';
+import {gameFormTypeTriple, gameFormTypeTripleHandler} from './../chunks/game-form-triple';
 
+const getQuestionFormAndHandler = (type) => {
+  const gameInterface = {};
+
+  let formElement = null;
+  let formHandler = null;
+
+  switch (type) {
+    case questionTypes.SIGNLE:
+      formElement = gameFormTypeSingle;
+      formHandler = gameFormTypeSingleHandler;
+      break;
+    case questionTypes.COUPLE:
+      formElement = gameFormTypeCouple;
+      formHandler = gameFormTypeCoupleHandler;
+      break;
+    case questionTypes.TRIPLE:
+      formElement = gameFormTypeTriple;
+      formHandler = gameFormTypeTripleHandler;
+      break;
+    default:
+      throw new Error(`Odd type of question: ${type}`);
+  }
+
+  gameInterface.form = formElement;
+  gameInterface.handler = formHandler;
+
+  return gameInterface;
+};
 
 export default class GameView extends AbstractView {
   constructor(data, question) {
@@ -15,7 +46,7 @@ export default class GameView extends AbstractView {
     return `
       <div class="game">
         <p class="game__task">${this._question.title}</p>
-        ${gameForm(this._question.type).form(this._question)}
+        ${getQuestionFormAndHandler(this._question.type).form(this._question)}
         <div class="stats">
           ${statsTemplate(this._gameData.answers)}
         </div>
@@ -27,7 +58,7 @@ export default class GameView extends AbstractView {
 
   bind(element) {
 
-    gameForm(this._question.type).handler(
+    getQuestionFormAndHandler(this._question.type).handler(
         element,
         {number: this._gameData.currentQuestion},
         this.onAnswer(this._gameData)
